Renamed base fitting problem
