+++ fitbenchmarking/parsing/parse_nist_data.py
@@ -28,12 +28,12 @@ from __future__ import (absolute_import, division, print_function)
 import os
 import numpy as np
 import re
-from parsing import fitbm_problem
+from parsing import base_fitting_problem
 
 from utils.logging_setup import logger
 
 
-class FittingProblem(fitbm_problem.BaseFittingProblem):
+class FittingProblem(base_fitting_problem.BaseFittingProblem):
     """
     Definition of the NIST problem class
